Extract IMDb GraphQL helper, format release dates, and fix send-as-photo
